in-processed

+++ src/services/auth.js
@@ -7,6 +7,13 @@ import {
   accessTokenLifetime,
   refreshTokenLifetime,
 } from '../constants/users.js';
+import jwt from 'jsonwebtoken';
+import { getEnvVar } from '../utils/getEnvVar.js';
+import { SMTP, TEMPLATES_DIR } from '../constants/index.js';
+import { sendEmail } from '../utils/sendEmail.js';
+import handlebars from 'handlebars';
+import path from 'node:path';
+import fs from 'node:fs/promises';
 
 const createSession = () => ({
   accessToken: randomBytes(30).toString('base64'),
@@ -18,7 +25,7 @@ const createSession = () => ({
 export const registerUser = async (payload) => {
   const { name, email, password } = payload;
   const user = await UsersCollection.findOne({ email });
-  console.log(payload);
+
   if (user) {
     throw createHttpError(409, 'User already exist');
   }
@@ -28,7 +35,9 @@ export const registerUser = async (payload) => {
     ...payload,
     password: hashPassword,
   });
+
   const { _id } = newUser;
+
   return {
     name,
     email,
@@ -45,6 +54,10 @@ export const loginUser = async ({ email, password }) => {
   const passwordCompare = await bcrypt.compare(password, user.password);
   if (!passwordCompare) {
     throw createHttpError(401, 'Password invalid');
+  }
+
+  if (!user.verify) {
+    throw createHttpError(401, 'Please verify your email');
   }
 
   await SessionCollection.deleteOne({ userId: user._id });
@@ -89,7 +102,65 @@ export const requestResetEmailToken = async (email) => {
   if (!user) {
     throw createHttpError(404, 'User not found');
   }
-  // create a token and save it to the user
+  const resetToken = jwt.sign(
+    {
+      sub: user._id,
+      email,
+    },
+    getEnvVar('JWT_SECRET'),
+    {
+      expiresIn: '5m',
+    },
+  );
+
+  const resetPasswordTemplatePath = path.join(
+    TEMPLATES_DIR,
+    'resetPasswordEmail.html',
+  );
+
+  const templateSource = (
+    await fs.readFile(resetPasswordTemplatePath)
+  ).toString();
+
+  const template = handlebars.compile(templateSource);
+
+  const html = template({
+    name: user.name,
+    link: `${getEnvVar('APP_DOMAIN')}/reset-pwd?token=${resetToken}`,
+  });
+
+  await sendEmail({
+    from: getEnvVar(SMTP.SMTP_FROM),
+    to: email,
+    subject: 'Reset your password',
+    html,
+  });
+};
+
+export const resetPassword = async (payload) => {
+  let entries;
+
+  try {
+    entries = jwt.verify(payload.token, getEnvVar('JWT_SECRET'));
+  } catch (error) {
+    if (error instanceof Error) throw createHttpError(401, error.message);
+    throw error;
+  }
+
+  const user = await UsersCollection.findOne({
+    email: entries.email,
+    _id: entries.sub,
+  });
+  if (!user) {
+    throw createHttpError(404, 'User not found');
+  }
+
+  const hashPassword = await bcrypt.hash(payload.password, 10);
+
+  await UsersCollection.updateOne(
+    { _id: user._id },
+    { password: hashPassword },
+  );
 };
 
 export const getUser = (filter) => UsersCollection.findOne(filter);
